LayoutTests/fast/profiler/dead-time.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

console.profile("Dead time in profile.");
function startTest()
{
    setTimeout("endTest()", 1234);
}

</script>
</head>

<body onload="startTest()">
This page's JavaScript Has some dead time in the profile.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be a large amount of the % spent in 
(non-JavaScript).
<div id="output"></div>
</body>
</html>
