LayoutTests/fast/profiler/dead-time-expected.txt

CONSOLE MESSAGE: line 59: Dead time in profile.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) dead-time.html 10
CONSOLE MESSAGE: line 75:   (program) dead-time.html 4
CONSOLE MESSAGE: line 75:   onload dead-time.html 19
CONSOLE MESSAGE: line 75:    startTest dead-time.html 12
CONSOLE MESSAGE: line 75:     setTimeout  0
CONSOLE MESSAGE: line 75:   (program) dead-time.html 1
CONSOLE MESSAGE: line 75:    endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript Has some dead time in the profile. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be a large amount of the % spent in (non-JavaScript).
